requirements.txt
isort
black
flake8
mypy
pyyaml
# pydantic
python-dotenv
streamlit
google-cloud-aiplatform
google-cloud-texttospeech